Replaced JPA Impls of evetrackerdata.services with spring managed repositories
Replaced usages inside the controllers with new Service interfaces that live inside the web submodule to let the structure match the functionality and to reduce code complexity. Reduces the interface size because the controllers don't need access to all CRUD operations. Created unit tests for the Impls of the new interfaces. Closes #60 Closes #61